[front] refactor: rename sandbox/pod_state.ts to db.ts, cover pod-state cold start in lifecycle tests
