+++ front/lib/api/sandbox/lifecycle.ts
@@ -10,7 +10,6 @@ import {
 } from "@app/lib/api/sandbox/instrumentation";
 import type { SandboxRuntimeOwner } from "@app/lib/api/sandbox/owner";
 import { podSandboxOnlyMounts } from "@app/lib/api/sandbox/pod_mounts";
-import { setupPodStateOnColdStart } from "@app/lib/api/sandbox/pod_state";
 import { startTelemetry } from "@app/lib/api/sandbox/telemetry";
 import type { Authenticator } from "@app/lib/auth";
 import { ConversationSandboxAdapter } from "@app/lib/resources/conversation_sandbox_adapter";
@@ -125,30 +124,6 @@ async function ensureOwnerSandboxReady(
         if (refreshResult.isErr()) {
           status = "error";
           return refreshResult;
-        }
-      }
-
-      // Pod state bring-up must run strictly AFTER the mounts (restore reads
-      // through the replica mount) and is awaited: `invoke` awaits
-      // ensurePodSandboxReady, which is what guarantees no function runs
-      // before the restore + daemon start completed. Only runs for pod
-      // owners.
-      if (freshlyCreated && runtimeOwner.kind === "pod") {
-        const podStateResult = await setupPodStateOnColdStart(auth, sandbox);
-        if (podStateResult.isErr()) {
-          // status=running was already committed by ensureActive, and this
-          // block only runs when freshlyCreated — a plain Err would make the
-          // NEXT call take the warm path and happily serve a half-initialized
-          // sandbox (unrestored databases, no litestream daemon). Request a
-          // kill instead: ensureActive's kill-requested branch destroys and
-          // recreates on the next access, re-running this setup from scratch.
-          logger.error(
-            { err: podStateResult.error, sandboxId: sandbox.sId },
-            "Pod state cold start failed — requesting sandbox kill so the next access recreates it."
-          );
-          await sandbox.requestKill();
-          status = "error";
-          return podStateResult;
         }
       }
