+++ front/lib/api/sandbox/pod_mounts.ts
@@ -1,6 +1,6 @@
 import type { SandboxOnlyMount } from "@app/lib/api/file_system/types";
 import { getPodSandboxFunctionsMountPoint } from "@app/lib/api/files/mount_path";
-import { POD_STATE_REPLICA_MOUNT_POINT } from "@app/lib/api/sandbox/pod_state";
+import { POD_STATE_REPLICA_MOUNT_POINT } from "@app/lib/api/sandbox/db";
 import type { SpaceResource } from "@app/lib/resources/space_resource";
 
 /**
